backend/requirements.txt
flask
flask-cors
torch
numpy
Pillow
tqdm
git+https://github.com/openai/CLIP.git
werkzeug
